between_app/requirements.txt
# requirements.txt

asgiref==3.8.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-database-url==2.2.0
Django==5.1
django-allauth==64.1.0
django-bootstrap5==24.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
python-decouple==3.8
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
resend==2.2.0
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
